feat: restructure and merge circuit into zone if unambiguously
